requirements.txt
google-genai==1.16.1
pillow==11.2.1
pytest==8.3.5
PyYAML==6.0.2
